% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPlots.R
\name{getPlots}
\alias{getPlots}
\title{getPlots: get Plot data}
\usage{
getPlots(
  plot_type = "VIBIplotID",
  survey_type = "all",
  hgm_class = "all",
  dom_veg1 = "all",
  plotID = "all",
  intens_mods = 1:4
)
}
\arguments{
\item{plot_type}{Filter on plot type. Options are "all", "VIBIplotID", (default), "oramID",
"wellID", and "wetlndID". Can choose multiple options.}

\item{survey_type}{Filter on survey type of plots. Options are "all" (default), "reference",
"survey", "survey, womc", "womc", and "womc, reference". Can choose multiple options.}

\item{hgm_class}{Filter on HGM class. Options are "all" (default), "Depression",
"Impoundment", "Riverine", "Slope". Can choose multiple options.}

\item{dom_veg1}{Filter on level 1 dominant vegetation type. Options are "all" (default), "emergent",
"forest", "shrub". Can choose multiple options.}

\item{plotID}{Quoted string. Default is 'all'. If specified will return data for only plots specified.
Can choose multiple plots. Based on FeatureID in database.}

\item{intens_mods}{Filter on total number of intensive modules. Ranges from 1 to 4. Can select multiple.
Default is 1:4 (all).}
}
\value{
Returns a data frame of plot data
}
\description{
This function filters plot-level location data data by plot and plot types.
}
\examples{
\dontrun{
# run first
importData()

# return all plots
plots <- getPlots()

# return only reference sites
ref <- getPlots(survey_type = c("reference", "womc, reference"))

# return only wetlands of management concern
womc <- getPlots(survey_type = c("womc", "womc, reference", "survey, womc"))

# return only depressional wetlands
depr <- getPlots(hgm_class = "Depression")

# return only forested vegetation types
forest <- getPlots(dom_veg1 = "forest")

# return non-forested vegetation types
nonfor <- getPlots(dom_veg1 = c("shrub", "emergent"))

# return plots for subset of plots
plots <- getPlots(plotID = c("1007", "1017", "1043"))
}

}
