% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joinVIBI.R
\name{joinVIBI}
\alias{joinVIBI}
\title{joinVIBI: join module-level VIBI metrics}
\usage{
joinVIBI(
  years = 2008:as.numeric(format(Sys.Date(), format = "\%Y")),
  survey_type = "all",
  hgm_class = "all",
  dom_veg1 = "all",
  plotID = "all",
  region = "NCNE"
)
}
\arguments{
\item{years}{Numeric. Filter on years of survey. Default is all.}

\item{survey_type}{Filter on survey type of plots. Options are "all" (default), "reference",
"survey", "survey, womc", "womc", and "womc, reference". Can choose multiple options.}

\item{hgm_class}{Filter on HGM class. Options are "all" (default), "Depression",
"Impoundment", "Riverine", "Slope". Can choose multiple options.}

\item{dom_veg1}{Filter on level 1 dominant vegetation type. Options are "all" (default), "Emergent",
"Forest", "Shrub". Can choose multiple options.}

\item{plotID}{Quoted string. Default is 'all'. If specified will return data for only plots specified.
Can choose multiple plots. Based on FeatureID in database.}

\item{region}{Quoted string. Default is "NCNE". Specifies the Army Corps Region for OH. Options are "EMP", "MW", and "NCNE".}
}
\value{
Returns a data frame of VIBI calculations for each module of each plot
}
\description{
This function summarizes module-level VIBI and filters by plot, year, and plot types. Note
that the Biomass metric assumes the aera sampled is always 0.01m2. Metric calculations and thresholds
follow INTEGRATED WETLAND ASSESSMENT PROGRAM Part 9: Field Manual for the Vegetation Index of
Biotic Integrity for Wetlands v. 1.5. Pages 17 - 20 and Table 2 were most useful. For plot-level average
VIBI scores, use sumVIBI. Currently, if there's only a biomass record for a given year, those records are
dropped.
}
\examples{
\dontrun{
# run first
importData()

 # ADD EXAMPLES

}

}
