% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joinVIBI_module.R
\name{joinVIBI_module}
\alias{joinVIBI_module}
\title{joinVIBI_module: join module-level VIBI metrics}
\usage{
joinVIBI_module(
  years = 2008:as.numeric(format(Sys.Date(), format = "\%Y")),
  survey_type = "all",
  hgm_class = "all",
  dom_veg1 = "all",
  plotID = "all",
  region = "NCNE",
  intens_mods = 1:4
)
}
\arguments{
\item{years}{Numeric. Filter on years of survey. Default is all.}

\item{survey_type}{Filter on survey type of plots. Options are "all" (default), "reference",
"survey", "survey, womc", "womc", and "womc, reference". Can choose multiple options.}

\item{hgm_class}{Filter on HGM class. Options are "all" (default), "Depression",
"Impoundment", "Riverine", "Slope". Can choose multiple options.}

\item{dom_veg1}{Filter on level 1 dominant vegetation type. Options are "all" (default), "emergent",
"forest", "shrub". Can choose multiple options. Note that emergent-coastal was not included as an option
in this VIBI, but wouldn't take much to add as an option.}

\item{plotID}{Quoted string. Default is 'all'. If specified will return data for only plots specified.
Can choose multiple plots. Based on FeatureID in database.}

\item{region}{Quoted string. Default is "NCNE". Specifies the Army Corps Region for OH. Options are "EMP", "MW", and "NCNE".}

\item{intens_mods}{Filter on total number of intensive modules. Ranges from 1 to 4. Can select multiple.
Default is 1:4 (all).}
}
\value{
Returns a data frame of VIBI calculations for each module of each plot
}
\description{
This function summarizes module-level VIBI and filters by plot, year, and plot types. This
function will return a VIBI score for each module within a site, and allows for more equal comparisons
among sites with fewer than 4 modules. Averaging the module scores within a site is also possible via sumVIBI.

NOTE 1: Because the results from sumVIBI are averages rather than sums (ie the VIBI spreadsheet), their
scores will be lower than other datasets that use the VIBI spreadsheet, which are straight sums.

NOTE 2: The Biomass metric assumes the area sampled is always 0.1m2. Currently, if there's only a biomass
record for a given year and no herb or woody data, those records are dropped.

NOTE 3: Metric calculations and thresholds follow INTEGRATED WETLAND ASSESSMENT PROGRAM Part 9: Field Manual
for the Vegetation Index of Biotic Integrity for Wetlands v. 1.5. Pages 17 - 20 and Table 2 were most useful.
}
\examples{
\dontrun{
# run first
importData()

# calculate module-level VIBI for all sites (default).
vibi <- joinVIBI_module()

write.csv(vibi, "./testing_scripts/vibimodule.csv", row.names=F)

# calculate module-level VIBI for plot 124 and year 2023
vibi124 <- joinVIBI_module(plotID = "124", years = 2023)

# calculate module-level VIBI for all sites sampled in 2023
vibi23 <- joinVIBI_module(years = 2023)

# calculate module-level VIBI for all emergent wetlands
vibi_emerg <- joinVIBI_module(dom_veg1 = "emergent")

# calculate module-level VIBI for depressional wetlands
vibi_dep <- joinVIBI_module(hgm_class = "Depression")


}

}
