% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joinVIBI_plot.R
\name{joinVIBI_plot}
\alias{joinVIBI_plot}
\title{joinVIBI_plot: join plot-level VIBI metrics}
\usage{
joinVIBI_plot(
  years = 2008:as.numeric(format(Sys.Date(), format = "\%Y")),
  survey_type = "all",
  hgm_class = "all",
  dom_veg1 = "all",
  plotID = "all",
  region = "NCNE"
)
}
\arguments{
\item{years}{Numeric. Filter on years of survey. Default is all.}

\item{survey_type}{Filter on survey type of plots. Options are "all" (default), "reference",
"survey", "survey, womc", "womc", and "womc, reference". Can choose multiple options.}

\item{hgm_class}{Filter on HGM class. Options are "all" (default), "Depression",
"Impoundment", "Riverine", "Slope". Can choose multiple options.}

\item{dom_veg1}{Filter on level 1 dominant vegetation type. Options are "all" (default), "emergent",
"forest", "shrub". Can choose multiple options. Note that emergent-coastal was not included as an option
in this VIBI, but wouldn't take much to add as an option.}

\item{plotID}{Quoted string. Default is 'all'. If specified will return data for only plots specified.
Can choose multiple plots. Based on FeatureID in database.}

\item{region}{Quoted string. Default is "NCNE". Specifies the Army Corps Region for OH. Options are "EMP", "MW", and "NCNE".}
}
\value{
Returns a data frame of VIBI calculations for each plot with 4 intensive modules
}
\description{
This function summarizes plot-level VIBI and filters by plot, year, and plot types.
NOTE 1: THIS FUNCTION ONLY WORKS WHERE 4 INTENSIVE MODULES ARE SAMPLED. This function is intended to provide
results that are comparable with other datasets that use the VIBI and are based on 4 intensive modules.
For module-level VIBI scores, use joinVIBI_module.

NOTE 2: The Biomass metric assumes the area sampled is always 0.01m2. Currently, if there's only a biomass
record for a given year and no herb or woody data, those records are dropped.

NOTE 3: Metric calculations and thresholds follow INTEGRATED WETLAND ASSESSMENT PROGRAM Part 9: Field Manual
for the Vegetation Index of Biotic Integrity for Wetlands v. 1.5. Pages 17 - 20 and Table 2 were most useful.
}
\examples{
\dontrun{
# run first
importData()

# calculate plot-level VIBI for all sites (default).
vibi <- joinVIBI_plot()
write.csv(vibi, "./testing_scripts/vibiplot.csv", row.names=F)

# calculate plot-level VIBI for plot 124 and year 2023
vibi124 <- joinVIBI_plot(plotID = "124", years = 2023)

# calculate plot-level VIBI for all sites sampled in 2023
vibi23 <- joinVIBI_plot(years = 2023)

# calculate plot-level VIBI for all emergent wetlands
vibi_emerg <- joinVIBI_plot(dom_veg1 = "emergent")

# calculate plot-level VIBI for depressional wetlands
vibi_dep <- joinVIBI_plot(hgm_class = "Depression")


}

}
