% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumVIBI.R
\name{sumVIBI}
\alias{sumVIBI}
\title{sumVIBI: summarize plot-level VIBI}
\usage{
sumVIBI(
  years = 2008:as.numeric(format(Sys.Date(), format = "\%Y")),
  survey_type = "all",
  hgm_class = "all",
  dom_veg1 = "all",
  plotID = "all",
  region = "NCNE"
)
}
\arguments{
\item{years}{Numeric. Filter on years of survey. Default is all.}

\item{survey_type}{Filter on survey type of plots. Options are "all" (default), "reference",
"survey", "survey, womc", "womc", and "womc, reference". Can choose multiple options.}

\item{hgm_class}{Filter on HGM class. Options are "all" (default), "Depression",
"Impoundment", "Riverine", "Slope". Can choose multiple options.}

\item{dom_veg1}{Filter on level 1 dominant vegetation type. Options are "all" (default), "Emergent",
"Forest", "Shrub". Can choose multiple options.}

\item{plotID}{Quoted string. Default is 'all'. If specified will return data for only plots specified.
Can choose multiple plots. Based on FeatureID in database.}

\item{region}{Quoted string. Default is "NCNE". Specifies the Army Corps Region for OH. Options are "EMP", "MW", and "NCNE".}

\item{nativity}{Quoted string. Filter on native status. Options are "all" (default), "adventive",
"cryptogeni", or "native". Can choose multiple statuses.}
}
\value{
Returns a data frame of VIBI calculations for each plot
}
\description{
This function summarizes plot-level VIBI and filters by plot, year, and plot types.
}
\examples{
\dontrun{
# run first
importData()

 # ADD EXAMPLES

}

}
