% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateSpeciesList.R
\name{updateSpeciesList}
\alias{updateSpeciesList}
\title{updateSpeciesList}
\usage{
updateSpeciesList(
  html_path = "./data/fqai_update_2018_08_13.html",
  save = TRUE,
  out_path = "./data/"
)
}
\arguments{
\item{html_path}{Quoted path, including file name, where the updated FQAI html table lives. Note that I tried
downloading the table directly from the epa.ohio.gov website, but the server did not allow it. You have
to download the latest html version of the FQAI list. The website housing the upated FQAI list is:
   https://epa.ohio.gov/divisions-and-offices/surface-water/reports-data/wetland-ecology.
The link to the html table of updated FQAI spreadsheet (2014) is:
   https://epa.ohio.gov/static/Portals/35/401/LU_Veg_Species.html.}

\item{save}{Logical. If TRUE (default), saves to the data folder in the package. If FALSE, only returns
object to global environment if named (see example).}

\item{out_path}{Quoted path you want to export joined species list to. Defaults to a data folder in your
working directory. Only needed if save = T.}
}
\value{
Returns a data frame of VIBI calculations for each plot
}
\description{
This function compares the tlu_WetlndSpeciesList in the HTLN database and the latest Ohio
EPA wetland species list with columns needed to rate the Ohio Vegetation IBI, and saves the table as
a csv to check against. This is mostly an internal function that may only be used whenever EPA updates
their list.
}
\examples{
\dontrun{
# run first
importData()

 # Write to file, but don't return output to global environment in R
 updateSpeciesList(html_path = "./data/fqai_update_2018_08_13.html", save = TRUE, out_path = "./data/")

 # Return spp_check to global env, but don't save to file
 spp_check <- updateSpeciesList(html_path = "./data/fqai_update_2018_08_13.html", save = F)

}

}
