% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBiomass.R
\name{getBiomass}
\alias{getBiomass}
\title{getBiomass: get biomass data}
\usage{
getBiomass(
  years = 2008:as.numeric(format(Sys.Date(), format = "\%Y")),
  survey_type = "all",
  hgm_class = "all",
  dom_veg1 = "all",
  plotID = "all",
  intens_mods = 1:4
)
}
\arguments{
\item{years}{Numeric. Filter on years of survey. Default is all.}

\item{survey_type}{Filter on survey type of plots. Options are "all" (default), "reference",
"survey", "survey, womc", "womc", and "womc, reference". Can choose multiple options.}

\item{hgm_class}{Filter on HGM class. Options are "all" (default), "Depression",
"Impoundment", "Riverine", "Slope". Can choose multiple options.}

\item{dom_veg1}{Filter on level 1 dominant vegetation type. Options are "all" (default), "emergent",
"forest", "shrub". Can choose multiple options.}

\item{plotID}{Quoted string. Default is 'all'. If specified will return data for only plots specified.
Can choose multiple plots. Based on FeatureID in database.}

\item{intens_mods}{Filter on total number of intensive modules. Ranges from 1 to 4. Can select multiple.
Default is 1:4 (all).}
}
\value{
Returns a data frame of biomass count data
}
\description{
This function filters biomass data data by plot, year, and plot types.
}
\examples{
\dontrun{
# run first
importData()

# return all biomass plant data
biomass <- getBiomass()

# return only 2023 data
biomass23 <- getBiomass(years = 2023)

# return 2010 and later
biomass10 <- getBiomass(years = 2010:2024)

# return only reference sites
ref <- getBiomass(survey_type = c("reference", "womc, reference"))

# return only wetlands of management concern
womc <- getBiomass(survey_type = c("womc", "womc, reference", "survey, womc"))

# return only depressional wetlands
depr <- getBiomass(hgm_class = "Depression")

return only forested vegetation types
forest <- getBiomass(dom_veg1 = "forest")

# return non-forested vegetation types
nonfor <- getBiomass(dom_veg1 = c("shrub", "emergent"))

# return biomass for subset of plots
biomass_plots <- getBiomass(plotID = c("1007", "1017", "1043"))

}

}
