% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importData().R
\name{importData}
\alias{importData}
\title{importData: Import tables from HTLN Wetlands database}
\usage{
importData(
  type = "DSN",
  odbc = "HTLNwetlands",
  filepath = NA,
  new_env = TRUE,
  data_type = "vibi"
)
}
\arguments{
\item{type}{Select how to import the database tables.
\describe{
\item{"DSN"}{Default. DSN database. If odbc is not specified, will default to HTLNwetlands.
Using this argument requires that you have a User DSN named HTLNwetlands that points to the
database containing the latest data.}
\item{"dbfile"}{A specified database containing the database tables. If selected,
must provide the database filepath in the filepath argument.}
\item{"csv"}{*NOT CURRENTLY ENABLED*. Imports the csv version of the data package views. If selected, must provide the
filepath for the csvs in the filepath argument. This option that does not require MS Access and
ODBC driver on your computer. }
\item{"zip"}{Imports the csv versions of the data package views, as a zipped file. If selected,
must provide the filepath and name of the zip file. This option that does not require MS Access and
ODBC driver on your computer.}
}}

\item{odbc}{DSN of the database when using type = DSN. If not specified will default to "HTLNwetlands",
which should represent the HTLN wetland database with the latest data.}

\item{filepath}{Quoted filepath where data package database (if type = "dbfile") or the csvs
(if type = "csv" or type = "zip") live.}

\item{new_env}{Logical. Specifies which environment to store views in. If \code{TRUE}(Default), stores
views in HTLNwetlands environment. If \code{FALSE}, stores views in global environment}

\item{data_type}{Select type of data to import
\describe{
\item{"vibi"}{Default. Import only VIBI-related tables. }
\item{"oram"}{*NOT CURRENTLY ENABLED* Import only ORAM-related tables.}
\item{"all"}{*NOT CURRENTLY ENABLED* Import all tables relevant for data package.}
}}
}
\value{
HTLN wetlands views in specified environment
}
\description{
This function imports tables from the Heartland Network database and compiles flat files
to be incorporated into the data package. You can either set a data source name (DSN) called "HTLNwetlands"
to import the data, or can specify a file name. Eventually this will be enabled to import the data package
csvs from IRMA.
}
\examples{
\dontrun{

# Import views using DSN
importData(type ='DSN', odbc = "HTLNwetlands") # this is the same as importData()

# Import views from specified database:
importData(type ='dbfile',
  filepath = 'C:/Users/KMMiller/OneDrive - DOI/MWR/HTLN_wetlands/VIBI/HTLNWetlands3.4.1.accdb')

}

}
