% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumVIBI.R
\name{sumVIBI}
\alias{sumVIBI}
\title{sumVIBI: summarize plot-level VIBI metrics}
\usage{
sumVIBI(
  years = 2008:as.numeric(format(Sys.Date(), format = "\%Y")),
  survey_type = "all",
  hgm_class = "all",
  dom_veg1 = "all",
  plotID = "all",
  region = "NCNE",
  intens_mods = 1:4
)
}
\arguments{
\item{years}{Numeric. Filter on years of survey. Default is all.}

\item{survey_type}{Filter on survey type of plots. Options are "all" (default), "reference",
"survey", "survey, womc", "womc", and "womc, reference". Can choose multiple options.}

\item{hgm_class}{Filter on HGM class. Options are "all" (default), "Depression",
"Impoundment", "Riverine", "Slope". Can choose multiple options.}

\item{dom_veg1}{Filter on level 1 dominant vegetation type. Options are "all" (default), "emergent",
"forest", "shrub". Can choose multiple options.}

\item{plotID}{Quoted string. Default is 'all'. If specified will return data for only plots specified.
Can choose multiple plots. Based on FeatureID in database.}

\item{region}{Quoted string. Default is "NCNE". Specifies the Army Corps Region for OH.
Options are "EMP", "MW", and "NCNE".}

\item{intens_mods}{Filter on total number of intensive modules. Ranges from 1 to 4. Can select multiple.
Default is 1:4 (all).}
}
\value{
Returns a data frame of average VIBI calculations for each plot
}
\description{
This function summarizes plot-level VIBI by averaging module-level VIBI scores, and
filters by plot, year, and plot types. Note that the Biomass metric assumes the area sampled is
always 0.01m2. Metric calculations and thresholds follow INTEGRATED WETLAND ASSESSMENT PROGRAM
Part 9: Field Manual for the Vegetation Index of Biotic Integrity for Wetlands v. 1.5.
Pages 17 - 20 and Table 2 were most useful. For plot-level VIBI scores that are comparable with
OH VIBI scores from other sites, use joinVIBI_plot(). Note however, that the joinVIBI_plot() only
rates sites with 4 intensive modules. This function allows sites with fewer than 4 modules to be
assessed on the same scale as sites with 4 modules. The VIBI scores with this function will be lower
than the original VIBI spreadsheet approach, because the VIBI spreadsheet sums across the 4 modules,
rather than averages. VIBI scores are applied after the module-level metrics are averaged to a plot-level value.
}
\examples{
\dontrun{
# run first
importData()

# average modules to plot-level VIBI for all sites (default).
vibi <- sumVIBI()

# average modules to plot-level VIBI for all sites sampled in 2023
vibi23 <- sumVIBI(years = 2023)

# average modules to plot-level VIBI for all emergent wetlands
vibi_emerg <- sumVIBI(dom_veg1 = "emergent")

# average modules to plot-level VIBI for depressional wetlands

vibi_dep <- sumVIBI(hgm_class = "Depression")


}

}
